requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.19.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.140
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
